chore: déplace les actions exporter et annoter dans le menu
