+++ front/src/components/collaborative/CollaborativeEditorArticleHeader.jsx
@@ -1,18 +1,11 @@
 import React from 'react'
-import { Eye, Printer } from 'lucide-react'
-import { Link } from 'react-router-dom'
 
 import useFetchData from '../../hooks/graphql.js'
-import { useModal } from '../../hooks/modal.js'
+
+import Loading from '../molecules/Loading.jsx'
 
 import { getArticleInfo } from '../Article.graphql'
 
-import Button from '../Button.jsx'
-
-import buttonStyles from '../button.module.scss'
-import Export from '../Export.jsx'
-import Modal from '../Modal.jsx'
-import Loading from '../molecules/Loading.jsx'
 import styles from './CollaborativeEditorArticleHeader.module.scss'
 
 /**
@@ -29,8 +22,6 @@ export default function CollaborativeEditorArticleHeader({ articleId }) {
       revalidateOnReconnect: false,
     }
   )
-  const exportModal = useModal()
-
   if (isLoading) {
     return <Loading />
   }
@@ -38,40 +29,6 @@ export default function CollaborativeEditorArticleHeader({ articleId }) {
   return (
     <header className={styles.header}>
       <h1 className={styles.title}>{data?.article?.title}</h1>
-      <div>
-        <Button
-          icon
-          title="Download a printable version"
-          onClick={() => exportModal.show()}
-        >
-          <Printer />
-        </Button>
-        <Link
-          to={`/article/${articleId}/preview`}
-          title="Preview (open a new window)"
-          target="_blank"
-          rel="noopener noreferrer"
-          className={buttonStyles.icon}
-        >
-          <Eye />
-        </Link>
-      </div>
-
-      <Modal
-        {...exportModal.bindings}
-        title={
-          <>
-            <Printer /> Export
-          </>
-        }
-      >
-        <Export
-          articleId={articleId}
-          name={data?.article?.title}
-          bib={data?.article?.workingVersion?.bibPreview}
-          onCancel={() => exportModal.close()}
-        />
-      </Modal>
     </header>
   )
 }
